fix(bf-5y0tyk): correct x-coordinate intersection rounding in Type3 rasterizer
Fixed intersection calculation to properly cast i32 to f64 before rounding:
- Old: edge.x.round() (doesn't compile - i32 has no round() method)
- New: (edge.x as f64).round() as i32

The intersections vector stores rounded x coordinates for each active edge,
used for fill span calculation in the scanline algorithm.

Acceptance criteria:
- Compute x = round(edge.x) for each active edge ✓
- Store intersections in Vec<i32> ✓
- Code compiles ✓

Closes bf-5y0tyk

Bead-Id: bf-5y0tyk

+++ crates/pdftract-core/src/font/type3_rasterizer.rs
@@ -1338,7 +1338,8 @@ impl<'a> RasterizerContext<'a> {
             aet.sort_by_key(|e| e.x);
 
             // Calculate intersection x coordinates for this scanline
-            let intersections: Vec<i32> = aet.iter().map(|edge| edge.x.round() as i32).collect();
+            // Compute x = round(edge.x) for each active edge
+            let intersections: Vec<i32> = aet.iter().map(|edge| (edge.x as f64).round() as i32).collect();
 
             // Fill between pairs of X positions (even-odd rule)
             for i in (0..intersections.len()).step_by(2) {
